+++ atlas/modules/commands/new_project.py
@@ -1,6 +1,7 @@
 from importlib import import_module
 import os
 import shutil
+import subprocess
 
 from atlas.modules.commands.base import BaseCommand, CommandError
 from atlas.conf import settings
@@ -15,6 +16,11 @@ class StartProjectCommand(BaseCommand):
 
     def add_arguments(self, parser):
         parser.add_argument("name", help="Name of the application or project.")
+        parser.add_argument(
+            "--auto-setup",
+            help="Name of Runner for which you want to auto-setup packages",
+            default="artillery"
+        )
 
     def handle(self, **options):
         project_name = options.pop("name")
@@ -27,21 +33,57 @@ class StartProjectCommand(BaseCommand):
             pass
         else:
             raise CommandError(
-                "{} conflicts with the name of an existing Python module and "
-                "cannot be used as a project name. Please try another name.".format(project_name)
+                f"{project_name} conflicts with the name of an existing Python module and "
+                "cannot be used as a project name. Please try another name."
             )
 
         target = os.path.join(os.getcwd(), project_name)
         if os.path.exists(target):
-            raise CommandError("{} directory already exists! Aborting".format(project_name))
+            overwrite = input(f"{project_name} directory already exists! Overwrite (y/N) ")
+            if overwrite.lower() == "y":
+                print(f"Removing existing {project_name}")
+                shutil.rmtree(target)
+            else:
+                print("Aborting!")
+                return
 
+        print(f"Creating {project_name}...")
         shutil.copytree(os.path.join(settings.BASE_DIR, settings.APP_TEMPLATE_DIRECTORY), target)
+
+        print("\nProject successfully created\n")
+
+        if options.get("auto_setup", "") == "artillery":
+            self.setup_artillery(target, project_name)
+
+    @staticmethod
+    def setup_artillery(target, project_name):
+        """
+        Setup packages for artillery
+        """
+
+        os.chdir(target)
+
+        print("Setting up packages for Artillery\n")
+
+        subprocess.run(["npm", "install", "-g", "artillery"])
+        subprocess.run(["npm", "install"])
+
+        print(success_message)
 
     @staticmethod
     def validate_name(name):
         # If it's not a valid directory name.
         if not name.isidentifier():
             raise CommandError(
-                "{} is not a valid project name. Please make sure the name is "
-                "a valid identifier.".format(name)
+                f"{name} is not a valid project name. Please make sure the name is a valid identifier."
             )
+
+
+success_message = """
+Congratulation, your project is successfully created! Follow these steps to configure your load test:
+1. Switch to project directory
+2. Run docker compose to install influxdb and grafana - `docker-compose -f docker/docker-compose.yml up`
+3. In the conf/ folder, configure conf.py, swagger.yaml, and profiles.yaml file
+4. Change settings.py as necessary
+5. Run `atlas dist`
+"""