+++ atlas/modules/transformer/commands/base.py
@@ -9,9 +9,12 @@ class TransformerBaseCommand(BaseCommand):
     VALID_CONVERTERS = ""
 
     def add_arguments(self, parser):
-        parser.add_argument("type", help="Load Tester Type which should be used. Valid types: {}".format(
-            self.VALID_CONVERTERS
-        ))
+        parser.add_argument(
+            "type",
+            nargs='?',      # This will make sure that we can pick argument from default or command line
+            default="artillery",
+            help=f"Valid build/dist types are: {self.VALID_CONVERTERS}"
+        )
 
     def handle(self, **options):
         raise NotImplementedError